+++ expr/expr_test.go
@@ -1272,17 +1272,6 @@ func TestEvalExpression(t *testing.T) {
 				[]float64{0, 1, 2, 3, 4, 5}, 1, now32)},
 		},
 		{
-			parser.NewExpr("absolute",
-
-				"metric1",
-			),
-			map[parser.MetricRequest][]*types.MetricData{
-				{"metric1", 0, 1}: {types.MakeMetricData("metric1", []float64{0, -1, 2, -3, 4, 5}, 1, now32)},
-			},
-			[]*types.MetricData{types.MakeMetricData("absolute(metric1)",
-				[]float64{0, 1, 2, 3, 4, 5}, 1, now32)},
-		},
-		{
 			parser.NewExpr("isNonNull",
 
 				"metric1",