+++ tests/helper.go
@@ -17,6 +17,19 @@ type FuncEvaluator struct {
 }
 
 func (evaluator *FuncEvaluator) EvalExpr(e parser.Expr, from, until int32, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
+	if e.IsName() {
+		return values[parser.MetricRequest{Metric: e.Target(), From: from, Until: until}], nil
+	} else if e.IsConst() {
+		p := types.MetricData{FetchResponse: pb.FetchResponse{Name: e.Target(), Values: []float64{e.FloatValue()}}}
+		return []*types.MetricData{&p}, nil
+	}
+	// evaluate the function
+
+	// all functions have arguments -- check we do too
+	if len(e.Args()) == 0 {
+		return nil, parser.ErrMissingArgument
+	}
+
 	return evaluator.eval(e, from, until, values)
 }