+++ CHANGES.md
@@ -18,7 +18,7 @@ CHANGELOG
  - [Improvement] graphiteWeb can query graphite-web 1.1.0+ for a list of supported functions and automatically do fallback to graphite in case user calls unimplemented function
  - [Fix] Fix typo in polyfit function description
  - [Code] Move MakeResponse to expr/types and rename it to MakeMetricData. This is sometimes useful outside of tests. Thx. to @borovskyav
- - [Code] Add helper package that will ease doing per-function tests. See `expr/functions/cairo/cairo_test.go` for an example
+ - [Code] Add helper package that will ease doing per-function tests. See `expr/functions/absolute/function_test.go` for an example
 
 **0.10.0.1**
  - [Fix] Autobuild scripts. Version bump to avoid retagging.